v21.04.09 aws s3 support

+++ fish-server/src/main/java/com/antsdb/saltedfish/cpp/Unsafe.java
@@ -182,7 +182,7 @@ public final class Unsafe {
         return unsafe.allocateMemory(size);
     }
 
-    public static boolean compareAndSwapLong(long addr, int expected, long value) {
+    public static boolean compareAndSwapLong(long addr, long expected, long value) {
         return unsafe.compareAndSwapLong(null, ix(addr), expected, value);
     }
